Change side mining payload BlockHeight type to uint32.

+++ ELA.CR.NewProposal/core/payloadsidemining.go
@@ -12,7 +12,7 @@ const SideMiningPayloadVersion byte = 0x00
 type PayloadSideMining struct {
 	SideBlockHash   Uint256
 	SideGenesisHash Uint256
-	BlockHeight     uint64
+	BlockHeight     uint32
 }
 
 func (a *PayloadSideMining) Data(version byte) []byte {
@@ -33,7 +33,7 @@ func (a *PayloadSideMining) Serialize(w io.Writer, version byte) error {
 	if err != nil {
 		return err
 	}
-	err = WriteUint64(w, a.BlockHeight)
+	err = WriteUint32(w, a.BlockHeight)
 	if err != nil {
 		return err
 	}
@@ -49,7 +49,7 @@ func (a *PayloadSideMining) Deserialize(r io.Reader, version byte) error {
 	if err != nil {
 		return err
 	}
-	height, err := ReadUint64(r)
+	height, err := ReadUint32(r)
 	if err != nil {
 		return err
 	}
